% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsSampler.R
\name{linRegGibbs}
\alias{linRegGibbs}
\title{Linear Regression Gibbs Sampler}
\usage{
linRegGibbs(X, testX, Y, testY, regVarPrior, lambdaSqPrior, numEpochs)
}
\arguments{
\item{X}{Design Matrix}

\item{testX}{Design Matrix}

\item{Y}{Outcome}

\item{testY}{Outcome}

\item{regVarPrior}{Scale hyperparameter for the Half-Cauchy prior for the regression variance.}

\item{lambdaSqPrior}{Scale hyperparameter for the Half-Cauchy prior for the regression coefficient prior variance.}

\item{numEpochs}{Number of epochs to run the Gibbs Sampler}
}
\value{
a list object consisting of five components
\describe{
\item{coefBeta}{a \eqn{p \times E} matrix of posterior samples for \eqn{p} the regression coefficientsm where \eqn{E} is the number of samples}
\item{lambdaSqScale}{vector of posterior samples for the auxiliary scale variable for the regression coefficient prior variance}
\item{lambdaSq}{vector of posterior samples for the regression coefficient prior variance}
\item{regVarScale}{vector of posterior samples for the auxiliary scale variable for the regression variance}
\item{regVar}{vector of posterior samples for the regression variance}
}
}
\description{
Estimates the posterior distribution of a Hierarchical Linear Model.
Places Half-Cauchy priors on the regression coefficient prior variance and the regression variance.
Introduces auxiliary scale parameters in order to sample a Half Cauchy variable
via sampling an inverse gamma - inverse gamma mixture using Gibbs sampling.
}
