% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsSampler.R
\name{linRegGibbsProcessed}
\alias{linRegGibbsProcessed}
\title{Gibbs Sampling With Multiple MCMC Chains Running In Parallel}
\usage{
linRegGibbsProcessed(
  X,
  testX,
  Y,
  testY,
  lambdaSqPrior,
  regVarPrior,
  numEpochs,
  numDiscard,
  numChains,
  numCores
)
}
\arguments{
\item{X}{Design Matrix}

\item{testX}{Design Matrix}

\item{Y}{Outcome}

\item{testY}{Outcome}

\item{lambdaSqPrior}{Scale hyperparameter for the Half-Cauchy prior for the regression coefficient prior variance.}

\item{regVarPrior}{Scale hyperparameter for the Half-Cauchy prior for the regression variance.}

\item{numEpochs}{Number of epochs to run the Gibbs Sampler}

\item{numDiscard}{Number of epochs to throw away as burn in}

\item{numChains}{Number of MCMC chains to run}
}
\value{
a list object consisting of three components
\describe{
\item{resMat}{matrix of posterior samples. Rows correspond to model parameters.}
\item{postMeanList}{vector of posterior means}
\item{RhatList}{vector of Rhat diagnostics}
}
}
\description{
Gibbs Sampling With Multiple MCMC Chains Running In Parallel
}
