% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsSampler.R
\name{Rhat}
\alias{Rhat}
\title{Calculate Rhat Convergence Diagnostics From MCMC Output}
\usage{
Rhat(samples, numChains)
}
\arguments{
\item{samples}{The number of samples in each chain.}

\item{numChains}{The number of chains.}
}
\value{
Rhat Convergence Diagnostics For A Single Model Parameter.
}
\description{
Calculate Rhat Convergence Diagnostics From MCMC Output
}
